config/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pillow>=10.0.0
opencv-python>=4.8.0
pytesseract>=0.3.10
pymongo>=4.6.0
motor>=3.3.0
requests>=2.31.0
numpy>=1.24.0
python-dotenv>=1.0.0
jinja2>=3.1.0
aiofiles>=0.23.0
aiohttp>=3.9.0